requirements.txt
# Core Databases
pymongo==4.6.0
neo4j==5.14.0
redis==5.0.1

# Utilities
python-dotenv==1.0.0
click==8.1.7
colorama==0.4.6
tabulate==0.9.0
rich==13.7.0
bcrypt==4.1.2
python-dateutil==2.8.2
pytz==2023.3.post1
pyfiglet==0.8.post1

# Development
pytest==7.4.3
black==23.12.1
flake8==7.0.0
PyJWT==2.8.0